allow for images with a static preset to scroll
The `touch-action: none` css property is set on the `.atlas` class indescriminantly, causing it to consume all touch events.  For cases where Atlas is intended to ‘own’ those events that’s fine. But, in the static mode, the window/document should own those events.

By exposing the render preset as a class, we can use that as a selector, and thus return the touch-events property to the default by setting it to inherit