adjust touch logic for static rendering
CanvasPanel can use a static renderer, but set interactive to `false`.  This works _ok_ in many cases, but falls down in a few.  This allows us to adjust the interactivity from the css level by moving the pointer-events logic into it’s own css class that’s  assigned based on the interactive prop