WIP: Rotation around arbitrary point
I’m still trying to figure out the exact position to start with (i.e. how to consistently get the center of the viewport, but I _think_ I found the basic logic to facilitate the rotation around the center.

The core of the issue is that when we do a translation, we translate the center of the thing to 0,0 and then rotate, and then translate back.  But, if we translate to 0,0 and then from there translate back not to the original point, but to the rotation from the center to that point, we should get what we want.  The maths aren’t too bad, it’s just getting the right info.